pyproject.toml
[project]
name = "prescience"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Oliver Normand", email = "oliver@normand.uk" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai[anthropic]",
    "exa-py",
    "valyu",
    "python-dotenv",
    "httpx",
    "rich",
    "logfire>=4.25.0",
    "pydantic-ai-slim[openai]>=1.63.0",
]

[project.scripts]
prescience = "prescience:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
